Create prefixPostfix.cpp
This progrm will convert the prefix notation from an input
file "Prefix.in" to postfix notation to an output file called "Postfix.out". The input file has
one expression per line. The expressions are made up of two types of symbols, single letter
operands and operators(+, -, *, /). There may be 0, 1 or more blanks between the symbols.
Each prefix expression from the input expression should be echo printed (to the screen). After the
program converts the input expression to its postfix equivalent, the postfix expression should be
printed to the output file, with an appropriate label.