Create fifoQueue.cpp
This program will reimplement the FIFO queue ADT using
a c++ int data type to contain the queue elements. In other words, one integer variable
is used as a queue that can store several integers. This program also prints the contents
of the queue using a member function printq. If the queue is empty, printq has no output at all.