Create writeCharRecur.cpp
This program uses recurssion to print character. Recursion is calling
a function from the same function.
The first recursive C++ function, writeLine that writes a character repeatedly 
to form a line of n characters.
The second recursive C++ function, writeBlock, that uses writeLine to write m 
lines of n characters each.
All input and output is from the main function.